Fix glob walking across cache-keys for relative paths
Being able to specify relative paths for cache keys is useful, for
example, when building a python project from maturin on a cargo
workspace (i.e. it's not just the local package to trigger rebuilds
from).

The previous tool, globwalker, cannot handle relative paths...

* https://github.com/Gilnaa/globwalk/issues/28

So it's back to plain old glob. Some assistance to redirect the
relative paths to the cwd is needed.